requirements.txt
django~=5.0.1
djangorestframework
django-filter

# Django environment variables
# https://django-environ.readthedocs.io/
django-environ

#Postgres Requirement
# sudo apt install python3-dev libpq-dev
#development
psycopg2-binary
#production ready
# psycopg2
